backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
motor==3.3.1  # Асинхронный драйвер MongoDB
pydantic==2.5.2
python-jose==3.3.0  # Для JWT токенов
passlib==1.7.4  # Для хеширования паролей
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2  # Для тестирования
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.0.1